+++ .eslintrc.json
@@ -3,13 +3,11 @@
     "browser": true,
     "es2021": true
   },
-
   "settings": {
     "import/resolver": {
       "typescript": {}
     }
   },
-
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "project": "./tsconfig.json",
@@ -26,23 +24,34 @@
     "airbnb",
     "airbnb-typescript"
   ],
-  "plugins": ["react", "@typescript-eslint", "prettier"],
+  "plugins": [
+    "react",
+    "@typescript-eslint",
+    "prettier"
+  ],
   "rules": {
     "prettier/prettier": "error",
-
-    "react/jsx-filename-extension": [1, { "extensions": [".tsx"] }],
+    "react/jsx-filename-extension": [
+      1,
+      {
+        "extensions": [
+          ".tsx"
+        ]
+      }
+    ],
     "react/react-in-jsx-scope": "off",
     "react/require-default-props": "off",
     "react/no-array-index-key": "off",
     "react/no-unescaped-entities": "off",
     "react/jsx-one-expression-per-line": "off",
-    
+    "react/jsx-wrap-multilines": "off",
     "import/no-extraneous-dependencies": "off",
     "import/order": "off",
-    "@typescript-eslint/no-use-before-define" : "off",
+    "@typescript-eslint/no-use-before-define": "off",
     "@typescript-eslint/no-explicit-any": "warn",
     "jsx-a11y/img-redundant-alt": "off",
     "react/button-has-type": "off",
-    "no-param-reassign": "off"
+    "no-param-reassign": "off",
+    "max-len": "off"
   }
 }
