+++ .prettierrc
@@ -1,4 +1,4 @@
 {
 	"semi": true,
-	"singleQuote": true
+	"singleQuote": true,
 }
