+++ package.json
@@ -22,7 +22,7 @@
     "framer-motion": "^10.16.1",
     "jotai": "^2.2.2",
     "next": "^13.3.4",
-    "postcss": "8.4.26",
+    "postcss": "^8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "3.3.3"
