+++ src/app/association/page.tsx
@@ -1,4 +1,3 @@
-import TimelineMark from '@/app/adherent/components/TimelineMark';
 import BecomeMember from '@/shared/components/BecomeMember.component';
 import { Avatar } from '@nextui-org/avatar';
 import {
@@ -7,6 +6,9 @@ import {
   IconMail,
   IconMapPin,
 } from '@tabler/icons-react';
+import { Image } from '@nextui-org/image';
+import Timeline from '@/shared/layout/Timeline.layout';
+import timelineEvents from './assets/timeline-events.json';
 
 type StaffMember = {
   name: string;
@@ -18,6 +20,9 @@ type StaffMember = {
   }[];
 };
 
+/**
+ * Page de présentation de l'association
+ */
 export default async function Page() {
   const staffMembers: StaffMember[] = [
     {
@@ -137,13 +142,13 @@ export default async function Page() {
   ];
 
   return (
-    <div className="flex flex-col gap-20 px-32 pt-12">
-      <div className="flex justify-between section">
-        <div className="w-5/12">
-          <p>Introduction</p>
-          <h1>Notre vision</h1>
+    <main>
+      <section className="section flex justify-between section gap-20 mt-28 flex-col lg:flex-row">
+        <div>
+          <small>Introduction</small>
+          <h1>Une association...</h1>
         </div>
-        <div className="flex flex-col gap-6 w-2/4">
+        <div className="flex flex-col gap-6 lg:w-1/2">
           <p>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce
             varius faucibus massa sollicitudin amet augue. Nibh metus a semper
@@ -165,49 +170,20 @@ export default async function Page() {
             arcu, tincidunt ultrices pellentesque.
           </p>
         </div>
-      </div>
-      <div className="flex justify-between section">
-        <div className="w-5/12">
-          <p>Chronologie</p>
-          <h1>Une association qui traverse les générations</h1>
+      </section>
+      <section className="section flex w-full justify-center items-center">
+        <Image
+          src="/assets/association/photo_groupe.png"
+          alt="Photo de groupe"
+        />
+      </section>
+      <div className="flex justify-between section gap-20 flex-col lg:flex-row">
+        <div>
+          <small>Chronologie</small>
+          <h1>... Qui traverse les générations</h1>
         </div>
-        <div className="flex flex-col w-2/4">
-          <TimelineMark
-            year={2005}
-            title="Création de l'association"
-            text="Un groupe d'étudiants de l'UBS de Vannes décide de créer une association pour sensibiliser le public à la protection des fonds marins. La Palme Verte est née."
-          />
-          <TimelineMark
-            year={2008}
-            title="L'équipe s'agrandit"
-            text="5 nouveaux membres rejoignent l'association."
-          />
-          <TimelineMark
-            year={2009}
-            title="Premier site web"
-            text="L'association se dote d'un site web pour étendre sa visibilité. Créé par Alain, il est toujours en ligne aujourd'hui."
-          />
-          <TimelineMark
-            year={2015}
-            title="Plus de 20 adhérents"
-            text="L'association franchit le cap des 20 adhérents. Elle est désormais reconnue dans le milieu associatif local."
-          />
-          <div className="flex h-44">
-            <div className="w-1/12">
-              {/* <div className="rounded-full bg-black"></div> */}
-              <svg>
-                <circle cx="20" cy="20" r="20" fill="green" />
-              </svg>
-            </div>
-            <div className="flex flex-col gap-2">
-              <p>2023</p>
-              <h1 className="text-2xl">Refonte graphique</h1>
-              <p>
-                La Palme Verte fait peau neuve avec son nouveau logo et un site
-                internet réimmaginé.
-              </p>
-            </div>
-          </div>
+        <div className="flex flex-col lg:w-1/2">
+          <Timeline events={timelineEvents} />
         </div>
       </div>
       <div className="flex justify-between section">
@@ -318,6 +294,6 @@ export default async function Page() {
           />
         </div>
       </div>
-    </div>
+    </main>
   );
 }
